Create, delete & update foodproduct now works

+++ public/supplier.php
@@ -9,11 +9,11 @@
 
 	$admId = $_SESSION['id'];
 	$privilege = $_SESSION['privilege'];
-	if($privilege == 2){
+	//if($privilege == 2){
 		$products = json_decode(file_get_contents("http://dev2-vyh.softwerk.se:8080/matkasseWS/rest/foodproduct/getbysupplier/".$admId),true);
-	}else{
+	/*}else{
 		header("Location: /index.php");	
-	}
+	}*/
 		
 ?>
 
@@ -30,7 +30,7 @@
                 echo "<div><h4 hidden>".$product["id"]."</h4></div>";
                 echo "<div><h4>".$product["name"]."</h4></div>";
                 echo "<a class='btn btn-default adm' href='/edit_supplier.php?id=".$product['id']."&name=".$product['name']."&cat_id=".$product["categoryID"]."'>Redigera</a>";
-                echo "<a class='btn btn-danger adm' href='/run.php?id=".$product['id']."&func=adm_supp_delete'>Ta bort</a></br>";
+                echo "<a class='btn btn-danger adm' href='/run.php?id=".$product['id']."&func=adm_supp_delete' onclick='return confirm(\"Är du säker på att du vill ta bort grossisten?\")'>Ta bort</a></br>";
             }
         ?>
      	<div><h3><a href='new_product.php'>Ny Vara</a></h3></div>
